backend/requirements.txt
fastapi
grpcio
grpcio-tools
uvicorn
Pillow
python-multipart
asyncio